pyproject.toml
[tool.poetry]
name = "fantasme"
version = "0.1.0"
description = ""
authors = ["pavlodubovyi <133531274+pavlodubovyi@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
pillow = "^11.1.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
